server/requirements/base.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
